harness/requirements.txt
openai>=1.50.0
PyYAML==6.0.2
